Initial unit test setup, route and template cleanup

+++ routes/html_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, request, redirect, url_for
+from flask import Blueprint, render_template, request, redirect, url_for, abort
 from db import db
 from models import User, Task
 
@@ -27,8 +27,11 @@ def tasks():
     return render_template("tasks.html", tasks=tasks)
 
 
-@html_routes_bp.route("/tasks/create", methods=["GET","POST"])
+@html_routes_bp.route("/tasks/create", methods=["GET", "POST"])
 def create_tasks():
+    """
+    Creates a task object using the POST request body
+    """
     if request.method == "POST":
         task_title = request.form["title"]
         task_description = request.form["description"]
@@ -43,7 +46,9 @@ def delete_task(task_id):
     """
     Deletes a task from the database based on the provided task ID
     """
-    task = Task.query.get_or_404(task_id)
+    task = db.session.get(Task, task_id)
+    if task is None:
+        abort(404)
     db.session.delete(task)
     db.session.commit()
     return redirect(url_for("html_routes.tasks"))
@@ -51,8 +56,13 @@ def delete_task(task_id):
 
 @html_routes_bp.route("/tasks/<int:task_id>/edit", methods=["POST"])
 def edit_task(task_id):
+    """
+    Edits a task from the database based on the provided task ID
+    """
     if request.method == "POST":
-        task = Task.query.get_or_404(task_id)
+        task = db.session.get(Task, task_id)
+        if task is None:
+            abort(404)
         task_description = request.form["description"]
         task_title = request.form["title"]
         task.name = task_title
